Add Domain Layer and useTweets and useTweet
