+++ projects/main/twitter/src/domain/useTweets.js
@@ -1,0 +1,34 @@
+import { useEffect, useState } from 'react';
+
+import { getTweets } from '../api/tweets';
+
+export default function useTweets() {
+  const [data, setData] = useState([]);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState('');
+
+  async function loadTweets() {
+    setLoading(true);
+    setError('');
+
+    try {
+      const response = await getTweets();
+
+      setData(response.data);
+    } catch (error) {
+      setError(error);
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  useEffect(() => {
+    loadTweets();
+  }, []);
+
+  return {
+    data,
+    loading,
+    error,
+  };
+}
